Component, image and video methods put back into CardListItem
