+++ App/Modules/index.js
@@ -6,6 +6,3 @@ export { default as Play, Drive, Beep } from './Play'
 export { default as Discover, CodeLab } from './Discover'
 export { Lessons, GetStarted } from './Learn'
 export { default as Lab } from './Lab'
-export { default as onVideoEnd, onVideoError, onVideoToggle } from './Video'
-export { default as getImageSize } from './Image'
-export { default as componentDidMount, componentWillMount } from './ComponentMount'
